begin description formatting
this is going to be a little hard cause we need the description
formatting to match on ios, android, and the web. maybe markdown
is best, if we can find a good react native markdown library?
dunno.

i'm inclined to do some markdown subset, like maybe just newlines,
bold, italic, maybe bullet points?